% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{check_DEqMS_parameter}
\alias{check_DEqMS_parameter}
\title{Helper function to check whether the DEqMS_PSMs_column is in the SummarizedExperiment object}
\usage{
check_DEqMS_parameter(se, DEqMS_PSMs_column)
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomics data set}

\item{DEqMS_PSMs_column}{String specifying which column name to use for DEqMS (default NULL)}
}
\value{
None
}
\description{
Helper function to check whether the DEqMS_PSMs_column is in the SummarizedExperiment object
}
\keyword{internal}
