% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{check_input_assays}
\alias{check_input_assays}
\title{Helper function to check whether all given assays are in SummarizedExperiment object}
\usage{
check_input_assays(se, ain)
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomics data set}

\item{ain}{Vector of strings which assay should be used as input (default NULL).
If NULL then all normalization of the se object are plotted next to each other.}
}
\value{
NULL if no methods in SummarizedExperiment object, else all available methods ready for visualization
}
\description{
Helper function to check whether all given assays are in SummarizedExperiment object
}
\keyword{internal}
