% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{check_plot_DE_parameters}
\alias{check_plot_DE_parameters}
\title{Helper function to check the parameters for plotting the DE results}
\usage{
check_plot_DE_parameters(de_res, ain, comparisons)
}
\arguments{
\item{de_res}{data table resulting of run_DE}

\item{ain}{String of normalization methods to visualize (must be valid normalization methods saved in de_res)}

\item{comparisons}{Vector of comparisons (must be valid comparisons saved in de_res)}
}
\value{
list of valid inputs for plotting functions
}
\description{
Helper function to check the parameters for plotting the DE results
}
\keyword{internal}
