% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{check_stats_spiked_DE_parameters}
\alias{check_stats_spiked_DE_parameters}
\title{Helper function to check the parameters for plotting the DE stats of spike-in data sets}
\usage{
check_stats_spiked_DE_parameters(stats, ain, comparisons)
}
\arguments{
\item{stats}{data table resulting of get_spiked_stats_DE}

\item{ain}{String of normalization methods to visualize (must be valid normalization methods saved in de_res)}

\item{comparisons}{Vector of comparisons (must be valid comparisons saved in de_res)}
}
\value{
list of valid inputs for plotting functions
}
\description{
Helper function to check the parameters for plotting the DE stats of spike-in data sets
}
\keyword{internal}
