% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download.R
\name{export_data}
\alias{export_data}
\title{Export the SummarizedExperiment object, the meta data, and the normalized data.}
\usage{
export_data(se, out_dir, ain = NULL)
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomics data set}

\item{out_dir}{Path of output directory}

\item{ain}{Vector of strings which assay should be downloaded (default NULL).
If NULL then all assays of the se object are saved.}
}
\value{
Nothing
}
\description{
Export the SummarizedExperiment object, the meta data, and the normalized data.
}
\examples{
data(tuberculosis_TMT_se)
\dontrun{export_data(tuberculosis_TMT_se, out_dir = "data/",
            ain = c("IRS_on_RobNorm", "IRS_on_Median"))}

}
