% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{expressToDT}
\alias{expressToDT}
\title{Helper function to transform an expression data frame to a data table}
\usage{
expressToDT(expr_data, column_names, row_names)
}
\arguments{
\item{expr_data}{Expression data frame containing the expression data}

\item{column_names}{Column names of the expression data}

\item{row_names}{Row names of the expression data}
}
\value{
Data table containing the expression data
}
\description{
Helper function to transform an expression data frame to a data table
}
\keyword{internal}
