% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEAnalysis.R
\name{extract_limma_DE}
\alias{extract_limma_DE}
\title{Extract the DE results from eBayes fit of perform_limma function.}
\usage{
extract_limma_DE(
  fit,
  comparisons,
  logFC = TRUE,
  logFC_up = 1,
  logFC_down = -1,
  p_adj = TRUE,
  alpha = 0.05
)
}
\arguments{
\item{fit}{eBayes object resulting from perform_limma method}

\item{comparisons}{Vector of comparisons that are performed in the DE analysis (from specify_comparisons method)}

\item{logFC}{Boolean specifying whether to apply a logFC threshold (TRUE) or not (FALSE)}

\item{logFC_up}{Upper log2 fold change threshold (dividing into up regulated)}

\item{logFC_down}{Lower log2 fold change threshold (dividing into down regulated)}

\item{p_adj}{Boolean specifying whether to apply a threshold on adjusted p-values (TRUE) or on raw p-values (FALSE)}

\item{alpha}{Threshold for adjusted p-values or p-values}
}
\value{
Data table with limma DE results
}
\description{
Extract the DE results from eBayes fit of perform_limma function.
}
