% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterProteins.R
\name{filter_out_complete_NA_proteins}
\alias{filter_out_complete_NA_proteins}
\title{Remove proteins with NAs in all samples}
\usage{
filter_out_complete_NA_proteins(se)
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomics data set}
}
\value{
filtered SummarizedExperiment object
}
\description{
Remove proteins with NAs in all samples
}
\examples{
data(tuberculosis_TMT_se)
tuberulosis_TMT_se <- filter_out_complete_NA_proteins(tuberculosis_TMT_se)

}
