% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateCompleteSE.R
\name{generate_complete_SE}
\alias{generate_complete_SE}
\title{Generate SummarizedExperiment object containing the samples of all assays in a single object (for overall PCA)}
\usage{
generate_complete_SE(se, ain = NULL)
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomics data set}

\item{ain}{String which assay should be used as input (default NULL)
If NULL then all normalization of the SummarizedExperiment object are plotted next to each other (except raw).}
}
\value{
SummarizedExperiment object with single assay containing the normalized samples of all assays
}
\description{
Generate SummarizedExperiment object containing the samples of all assays in a single object (for overall PCA)
}
\examples{
data(tuberculosis_TMT_se)
tuberculosis_TMT_se_all <- generate_complete_SE(tuberculosis_TMT_se,
                                                ain = NULL)
}
