% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{get_color_value}
\alias{get_color_value}
\title{Helper function to get correct value for coloration of plots (color_by parameter)}
\usage{
get_color_value(se, color_by)
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomics data set}

\item{color_by}{String specifying the column to color the samples (If NULL, the condition column of the SummarizedExperiment object is used. If "No", no color bar added.)}
}
\value{
String of column to color or NULL if no color should be applied
}
\description{
Helper function to get correct value for coloration of plots (color_by parameter)
}
\keyword{internal}
