% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationPlots.R
\name{get_complete_pca_dt}
\alias{get_complete_pca_dt}
\title{Function to get a long data table of all PCA1 and PCA2 values of all kind of normalization}
\usage{
get_complete_pca_dt(se, ain = NULL)
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomics data set}

\item{ain}{String which assay should be used as input (default NULL)
If NULL then all normalization of the SummarizedExperiment object are plotted next to each other (except raw).}
}
\value{
data table
}
\description{
Function to get a long data table of all PCA1 and PCA2 values of all kind of normalization
}
