% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{get_condition_value}
\alias{get_condition_value}
\title{Helper function to check the condition value}
\usage{
get_condition_value(se, condition)
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomics data set}

\item{condition}{column name of condition (if NULL, condition saved in SummarizedExperiment will be taken)}
}
\value{
String of column for condition
}
\description{
Helper function to check the condition value
}
\keyword{internal}
