% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{get_facet_value}
\alias{get_facet_value}
\title{Helper function to get correct value for faceting of plots (facet_by parameter)}
\usage{
get_facet_value(se, facet_by)
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomics data set}

\item{facet_by}{String specifying the column to facet the samples (If NULL or "No", no faceting is done.)}
}
\value{
String of column to facet or NULL if no faceting should be done
}
\description{
Helper function to get correct value for faceting of plots (facet_by parameter)
}
\keyword{internal}
