% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{get_label_value}
\alias{get_label_value}
\title{Helper function to get correct value for sample labeling of plots (label_by parameter)}
\usage{
get_label_value(se, label_by)
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomics data set}

\item{label_by}{String specifying the column to label the samples (If NULL, the labels column of the SummarizedExperiment object is used. If "No", no labeling of samples done.)}
}
\value{
String of column to label or NULL if no label should be applied
}
\description{
Helper function to get correct value for sample labeling of plots (label_by parameter)
}
\keyword{internal}
