% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetOverview.R
\name{get_NA_overview}
\alias{get_NA_overview}
\title{Function returning some values on the numbers of NA in the data}
\usage{
get_NA_overview(se, ain = "log2")
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomics data set}

\item{ain}{String which data type should be used (default raw)}
}
\value{
list with total amount of values in the data, amount of NA values, and the percentage of NAs
}
\description{
Function returning some values on the numbers of NA in the data
}
\examples{
data(tuberculosis_TMT_se)
get_NA_overview(tuberculosis_TMT_se, ain="log2")
}
