% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{get_normalization_methods}
\alias{get_normalization_methods}
\title{Function to return available normalization methods' identifier names}
\usage{
get_normalization_methods()
}
\value{
Vector of normalization methods
}
\description{
Function to return available normalization methods' identifier names
}
\examples{
get_normalization_methods()
  

}
