% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEAnalysis.R
\name{get_overview_DE}
\alias{get_overview_DE}
\title{Get overview table of DE results}
\usage{
get_overview_DE(de_res)
}
\arguments{
\item{de_res}{data table resulting of run_DE}
}
\value{
data table of numbers of DE proteins per comparison and per normalization method
}
\description{
Get overview table of DE results
}
\examples{
data(tuberculosis_TMT_de_res)
get_overview_DE(tuberculosis_TMT_de_res)

}
