% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{get_shape_value}
\alias{get_shape_value}
\title{Helper function to get correct value for shaping of plots (shape_by parameter)}
\usage{
get_shape_value(se, shape_by)
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomics data set}

\item{shape_by}{String specifying the column to shape the samples (If NULL or "No", no shaping is done.)}
}
\value{
String of column to shape or NULL if no shaping should be done
}
\description{
Helper function to get correct value for shaping of plots (shape_by parameter)
}
\keyword{internal}
