% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeDEPlots.R
\name{get_spiked_stats_DE}
\alias{get_spiked_stats_DE}
\title{Get performance metrics of DE results of spike-in data set.}
\usage{
get_spiked_stats_DE(se, de_res)
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomics data set}

\item{de_res}{data table resulting of run_DE}
}
\value{
data table with multiple performance metrics of the DE results
}
\description{
Get performance metrics of DE results of spike-in data set.
}
\examples{
data(spike_in_se)
data(spike_in_de_res)
stats <- get_spiked_stats_DE(spike_in_se, spike_in_de_res)

}
