% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationMethods.R
\name{loessCycNorm}
\alias{loessCycNorm}
\title{Cyclic Loess Normalization of limma}
\usage{
loessCycNorm(se, ain = "log2", aout = "LoessCyc", on_raw = FALSE)
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomic dataset}

\item{ain}{String which assay should be used as input}

\item{aout}{String which assay should be used to save normalized data}

\item{on_raw}{Boolean specifying whether normalization should be performed on raw or log2-scaled data}
}
\value{
SummarizedExperiment containing the loessCyc normalized data as assay (on log2 scale)
}
\description{
Two samples of the data are MA transformed and normalized at a time, and all pairs of samples are iterated through. Log2-scaled data should be taken as input (on_raw = FALSE).
}
\examples{
data(tuberculosis_TMT_se)
tuberculosis_TMT_se <- loessCycNorm(tuberculosis_TMT_se, ain = "log2",
                                      aout = "LoessCyc", on_raw = FALSE)

}
\seealso{
\code{\link{limma::normalizeCyclicLoess()}}
}
