% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetOverview.R
\name{plot_condition_overview}
\alias{plot_condition_overview}
\title{Barplot showing the number of samples per condition}
\usage{
plot_condition_overview(se, condition = NULL)
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomics data set}

\item{condition}{column name of condition (if NULL, condition saved in SummarizedExperiment will be taken)}
}
\value{
ggplot object
}
\description{
Barplot showing the number of samples per condition
}
\examples{
data(tuberculosis_TMT_se)
plot_condition_overview(tuberculosis_TMT_se, condition = NULL)
}
