% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeOverview.R
\name{plot_histogram_spiked}
\alias{plot_histogram_spiked}
\title{Plot histogram of the spike-in and background protein intensities per condition.}
\usage{
plot_histogram_spiked(se, condition = NULL)
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomics data set}

\item{condition}{column name of condition (if NULL, condition saved in SummarizedExperiment will be taken)}
}
\value{
ggplot object
}
\description{
Plot histogram of the spike-in and background protein intensities per condition.
}
\examples{
data(spike_in_se)
plot_histogram_spiked(spike_in_se, condition = NULL)

}
