% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterProteins.R
\name{plot_NA_density}
\alias{plot_NA_density}
\title{Plot the intensity distribution of proteins with and without NAs}
\usage{
plot_NA_density(se)
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomics data set}
}
\value{
ggplot object
}
\description{
Plot the intensity distribution of proteins with and without NAs
}
\examples{
data(tuberculosis_TMT_se)
plot_NA_density(tuberculosis_TMT_se)

}
