% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{readPRONE_example}
\alias{readPRONE_example}
\title{Helper function to read example data}
\usage{
readPRONE_example(path = NULL)
}
\arguments{
\item{path}{NULL to get all example data set files, otherwise specify the file name}
}
\value{
If path=NULL a character vector with the file names, otherwise the path to the specific file
}
\description{
Helper function to read example data
}
\examples{
readPRONE_example()
}
