% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterSamples.R
\name{remove_reference_samples}
\alias{remove_reference_samples}
\title{Remove reference samples of SummarizedExperiment object (reference samples specified during loading)}
\usage{
remove_reference_samples(se)
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomics data set}
}
\value{
filtered SummarizedExperiment object
}
\description{
Remove reference samples of SummarizedExperiment object (reference samples specified during loading)
}
\examples{
data(tuberculosis_TMT_se)
tuberculosis_TMT_se <- remove_reference_samples(tuberculosis_TMT_se)

}
