% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEAnalysis.R
\name{spectraCounteBayes_DEqMS}
\alias{spectraCounteBayes_DEqMS}
\title{Additional function of the DEqMS package}
\usage{
spectraCounteBayes_DEqMS(fit, coef_col)
}
\arguments{
\item{fit}{linear model from function perform_limma}

\item{coef_col}{an integer vector indicating the column(s) of fit$coefficients for which the function is to be performed. if not specified, all coefficients are used.}
}
\value{
list object
}
\description{
Additional function of the DEqMS package
}
