% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spike_in_de_res}
\alias{spike_in_de_res}
\title{Example data.table of DE results of of a spike-in proteomics data set}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 33470 rows and 10 columns.
}
\source{
Jürgen Cox, Marco Y. Hein, Christian A. Luber, Igor Paron, Nagarjuna Nagaraj, and Matthias Mann.Accurate Proteome-wide Label-free Quantification by Delayed Normalization and Maximal Peptide Ratio Extraction, Termed MaxLFQ. Molecular & Cellular Proteomics 13.9 (Sept. 2014), pp. 2513–2526. <https://doi.org/10.1074/mcp.M113.031591>.
}
\usage{
spike_in_de_res
}
\description{
A \code{data.table} containing the DE results of the spike_in_se data set (limma, logFC > 1, logFC < -1, p.adj < 0.05)
}
\keyword{datasets}
