% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tuberculosis_TMT_de_res}
\alias{tuberculosis_TMT_de_res}
\title{Example data.table of DE results of a real-world proteomics data set}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 9030 rows and 9 columns.
}
\source{
Biadglegne et al. Mycobacterium tuberculosis Affects Protein and Lipid Content of Circulating Exosomes in Infected Patients Depending on Tuberculosis Disease State. Biomedicines 10.4 (Mar. 2022), p. 783. doi: 10.3390/ biomedicines10040783.
}
\usage{
tuberculosis_TMT_de_res
}
\description{
A \code{data.table} containing the DE results of the tuberculosis_TMT_se data set (limma, logFC > 1, logFC < -1, p.adj < 0.05)
}
\keyword{datasets}
