% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterProteins.R
\name{plot_NA_frequency}
\alias{plot_NA_frequency}
\title{Plot protein identification overlap (x = identified in number of Samples, y=number of proteins)}
\usage{
plot_NA_frequency(se)
}
\arguments{
\item{se}{SummarizedExperiment containing all necessary information of the proteomics data set}
}
\value{
ggplot object
}
\description{
Plot protein identification overlap (x = identified in number of Samples, y=number of proteins)
}
\examples{
data(tuberculosis_TMT_se)
plot_NA_frequency(tuberculosis_TMT_se)

}
